fix: refactor api crate
