--- crates/beacon_api/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "ream-beacon-api"
-authors.workspace = true
-edition.workspace = true
-keywords.workspace = true
-license.workspace = true
-readme.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-enr.workspace = true
-discv5.workspace = true
-futures.workspace = true
-hex.workspace = true
-libp2p-identity.workspace = true
-libp2p-mplex.workspace = true
-libp2p.workspace = true
-serde.workspace = true
-serde_yaml.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-warp.workspace=true
-clap.workspace=true
-thiserror.workspace=true
-
-
-ream-config={workspace = true}
