--- crates/beacon_api/src/types/genesis.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug,Clone,Serialize,Deserialize)]
-pub struct GenesisData{
-    pub genesis_time:String,
-    pub genesis_validator_root:String,
-    pub genesis_fork_version:String,
-}
