--- crates/beacon_api/src/utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod error;
