+++ crates/rpc/Cargo.toml
@@ -9,4 +9,22 @@ repository.workspace = true
 rust-version.workspace = true
 version.workspace = true
 
+
 [dependencies]
+alloy-primitives.workspace = true
+anyhow.workspace = true
+enr.workspace = true
+discv5.workspace = true
+futures.workspace = true
+hex.workspace = true
+libp2p-identity.workspace = true
+libp2p-mplex.workspace = true
+libp2p.workspace = true
+serde.workspace = true
+serde_yaml.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+warp.workspace=true
+clap.workspace=true
+thiserror.workspace=true
