+++ crates/rpc/src/config.rs
@@ -1,4 +1,5 @@
 
+/// Server Config.
 #[derive(Debug)]
 pub struct ServerConfig {
     pub http_port: usize,
@@ -8,15 +9,11 @@ pub struct ServerConfig {
 
 impl ServerConfig {
     /// Creates a new instance from CLI arguments
-    pub fn from_args(
-        http_port: usize,
-        http_address: String,
-        http_allow_origin: bool,
-    ) -> Self {
+    pub fn from_args(http_port: usize, http_address: String, http_allow_origin: bool) -> Self {
         Self {
             http_port: http_port,
             http_address: http_address,
-            http_allow_origin
+            http_allow_origin,
         }
     }
