+++ crates/rpc/src/utils/error.rs
