+++ crates/rpc/src/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod chain;
+pub mod error;
